% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAndVisualizeCorrelatedCells.R
\name{findCorrelatedCells}
\alias{findCorrelatedCells}
\title{Computes the Pearson correlation between the gene expression of a
comparator dataframe and each spot in 10X Visium spatial data. Visualize
the pearson correlation significance of each Visium spot.}
\usage{
findCorrelatedCells(
  comparator,
  visiumData,
  assay = "Spatial",
  slot = "data",
  topGenes = 25
)
}
\arguments{
\item{comparator}{A dataframe of gene expression data, with one column
containing expression values. The rownames of the dataframe has to be set to
the gene names corresponding to the expression values}

\item{visiumData}{A 10X Visium spatial dataset loaded as a Seurat object}

\item{assay}{The assay in the Seurat object to use for correlation with
comparator, uses the Spatial assay by default}

\item{slot}{The specific assay data to use, uses normalized data matrix
(i.e. data slot) by default.}

\item{topGenes}{The number of top genes in the comparator to use for pearson
correlation with spatial data, by default 25. If the comparator has less than
25 genes, by default all genes in comparator are used for correlation with
spatial data.}
}
\value{
The inputted spatialDataset Seurat object modified metadata fields
"pearson.pvalue" and "pearson.estimate". "pearson.pvalue" stores the coefficient
of the pearson correlation between the comparator and the corresponding
Visium spot. "pearson.pvalue" that corresponds to the p-value of the pearson
correlation between the gene expression of the comparator and each Visium
spot

examples will be added
}
\description{
A function that computes the pearson correlation of the gene expression of
the top expressed genes in a comparator dataframe and each spot in 10X
Visium spatial dataset. The alternative hypothesis is that there is a
positive correlation between the spot and comparator dataframe.
}
\references{
Add references here
}
