% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPearsonCorrelateCells.R
\name{runPearsonCorrelateCells}
\alias{runPearsonCorrelateCells}
\title{Launch Shiny App for pearsonCorrelateCells}
\usage{
runPearsonCorrelateCells()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for pearsonCorrelateCells.
The purpose of this app will be added
}
\examples{
\dontrun{

pearsonCorrelateCells::runPearsonCorrelateCells()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
