% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAndVisualizeCorrelatedCells.R
\name{visualizeCells}
\alias{visualizeCells}
\title{A function that visualizes the pearson correlation significance of each
Visium spot overlayed on the Visium image}
\usage{
visualizeCells(visiumData)
}
\arguments{
\item{visiumData}{A 10X Visium spatial dataset loaded as a Seurat object that
has been processed using the findCorrelatedCells function}
}
\value{
A plot showing the pearson correlation significance of each
Visium spot overlayed on the Visium image

examples will be added
}
\description{
A function that visualizes the pearson correlation significance of each
Visium spot overlayed on the Visium image
}
\references{
Add references here
}
