% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAndVisualizeCorrelatedCells.R
\name{visualizeCells}
\alias{visualizeCells}
\title{Visualize the significance computed Pearson correlations in each visium cell
spot}
\usage{
visualizeCells(visiumData)
}
\arguments{
\item{visiumData}{A 10X Visium spatial dataset loaded as a Seurat object that
has been processed using the findCorrelatedCells function}
}
\value{
A plot showing the pearson correlation significance of each
Visium spot overlayed on the Visium image
}
\description{
A function that visualizes the pearson correlation significance of each
Visium spot, overlayed on the Visium image
}
\examples{
\dontrun{
# Using the 10X Genomics Visium Mouse Brain Dataset from SeuratData
SeuratData::InstallData("stxBrain")
brain <- SeuratData::LoadData("stxBrain", type = "posterior1")
# Using topOligoGenes dataset in package
brainWithPearson <- findCorrelatedCells(comparator = topOligoGenes,
visiumData = brain)
visualizeCells(brainWithPearson)
}

}
\references{
Add references here
}
